+++ lumino/__init__.py
@@ -3,10 +3,10 @@ import json
 
 _conf_dict = None
 
-def lumen():
-    with open('./lumino.conf', 'r') as file:
+def lumino():
+    with open('./lumino/lumino.conf', 'r') as file:
         _conf_dict = json.load(file)
-        print(json.dumps(conf_dict))
+        print(json.dumps(_conf_dict))
     return _conf_dict
     
 def version():
