+++ lumino/lumino.conf
@@ -1,8 +1,4 @@
 {
-  "12hour": true,
-  "24hour": false,
-  "back": true,
-  "color": "000000",
-  "dots": true,
-  "ntp": true
+  "wap": "decatur",
+  "wap-passwd": "a069a06f6d"
 }
