+++ templates/index.html
@@ -14,12 +14,12 @@
     <div class="h-64">
       <div class="p-4 m-4 bg-green-600">
         <img style="display:inline-block" src="static/icons/vortex.png" alt="VIL" width="150" height="150">
-        <div style="margin-left:1%;display:inline-block;font-family:DymaxionScript;font-size:72px" class="text-2xl font-bold text-white">Lumen&nbsp;&nbsp;Display</div>
+        <div style="margin-left:1%;display:inline-block;font-family:DymaxionScript;font-size:72px" class="text-2xl font-bold text-white">Lumino&nbsp;&nbsp;Display</div>
         <div style="font-family:DymaxionScript;font-size:20px" class="text-2xl font-bold text-black">Vortex&nbsp;&nbsp;Instrument&nbsp;&nbsp;Labs</div>
       </div>
       <div class="p-4 m-4 bg-green-300 text-green-900 h-auto">
         <div>
-          <h1 style="display:inline"> Lumen: </h1>
+          <h1 style="display:inline"> Lumino: </h1>
           <div style="display:inline">{{ date }}</div>
         </div>
         <br />
@@ -40,8 +40,12 @@
           <div style="display:inline">{{ wap }}</div>
         </div>
         <div>
+          <h2 style="display:inline" align=left>clock skew: </h2>
+          <div style="display:inline">{{ skew }}</div>
+        </div>
+        <div>
           <h2 style="display:inline" align=left>configuration: </h2>
-          <div style="display:inline">{{ conf_dict }}</div>
+          <div style="display:inline">{{ conf }}</div>
         </div>
       </div>
       <div class="p-4 m-4 bg-green-300 text-green-900 h-auto">
@@ -69,7 +73,7 @@
         <br />
         <div>
           <h2 style="display:inline" align=left>version: </h2>
-          <div style="display:inline" id="version">{{ NCS31X }}</div>
+          <div style="display:inline" id="version">{{ 'NCS31X' }}</div>
         </div>
         <div>
           <h2 style="display:inline" align=left>configuration: </h2>
